Remove some "float" cast. Use powf and sqrtf instead of the double version.
git-svn-id: svn://maxr.org/MAX/trunk@3144 c3bd5e88-ca0b-48cf-b972-da51c77c4c1b

+++ attackJobs.cpp
@@ -257,10 +257,10 @@ void cServerAttackJob::sendFireCommand()
 
 	float dx = (float) targetX - (float) agressorX;
 	float dy = (float) - (targetY - agressorY);
-	float r = sqrt (dx * dx + dy * dy);
+	float r = sqrtf (dx * dx + dy * dy);
 
 	int fireDir = unit->dir;
-	if (r <= 0.001)
+	if (r <= 0.001f)
 	{
 		//do not rotate agressor
 	}
